+++ packages/complex-rubric/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric-configure@2.2.3...@pie-element/complex-rubric-configure@2.2.4) (2024-03-26)
+
+**Note:** Version bump only for package @pie-element/complex-rubric-configure
+
+
+
+
+
 ## [2.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric-configure@2.2.2...@pie-element/complex-rubric-configure@2.2.3) (2024-03-22)
 
 **Note:** Version bump only for package @pie-element/complex-rubric-configure