+++ packages/drawing-response/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.7.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response-configure@7.7.0...@pie-element/drawing-response-configure@7.7.1) (2024-03-26)
+
+**Note:** Version bump only for package @pie-element/drawing-response-configure
+
+
+
+
+
 # [7.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response-configure@7.6.0...@pie-element/drawing-response-configure@7.7.0) (2024-03-20)