+++ packages/matrix/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.20.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.20.0...@pie-element/matrix-configure@0.20.1) (2024-03-26)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
 # [0.20.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.19.0...@pie-element/matrix-configure@0.20.0) (2024-03-20)