+++ packages/number-line/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/number-line-configure@7.4.0...@pie-element/number-line-configure@7.4.1) (2024-03-26)
+
+**Note:** Version bump only for package @pie-element/number-line-configure
+
+
+
+
+
 # [7.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/number-line-configure@7.3.0...@pie-element/number-line-configure@7.4.0) (2024-03-20)